refactor schema IT
